Fix various unicode conversion issues.
Templatize FileUtils::Tokenize so that it can be used with
std::wstring. Add functions for converting to the current
system codepage for all platforms. These are no-ops on
Linux and OS X.

Add stub functions which do 'dumb' conversions from std::string
to std::wstring on Linux and OS X. These should eventually be
replaced by functions that actually do the conversion.