More intelligent handling of Ruby symbols.
A Ruby symbol transferred to Kroll will be converted to
a string. Ruby hashes may be hashes of symbols. If the has
has more than half of its keys as symbols, new keys will
be symbols. If a string key is used which matches an
existing symbol in the hash, the key will be treated as
symbolic as well.

[#57 state:resolved]